# src/api/docs/openapi.yaml
openapi: 3.0.1
info:
  title: EPS API
  description: >-
    This is the API which supports the web and mobile applications for the
    Extended Prototype System (EPS).
  contact:
    email: jclements@ascend.network
  version: 0.29.0
...
paths:
...
  /roles:
    get:
      tags:
        - roles
      summary: Returns all roles in the system
      responses:
        200:
          description: Success operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'

    post:
      tags:
        - roles
      summary: Add role to system
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRole'
      responses:
        200:
          description: Success operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
  /roles/<int:role_id>:
    get:
      tags:
        - roles
      summary: Get an individual role
      responses:
        200:
          description: Success operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'

    put:
      tags:
        - roles
      summary: Update role information
      description: |
        This endpoint updates a role's information.

        If a role's permissions are updated such that it used to have the
        `ASSIGNABLE_TO_CASE` permission and now will not:
          * cases assigned to users with the role will be assigned
          * users with the admin role's will be notified about the affected cases.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRole'
      responses:
        200:
          description: Success operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'

    delete:
      tags:
        - roles
      summary: Deletes a role
      description: |
        This endpoint deletes a role from the system.

        The system provided Admin role cannot be deleted. Attempts to delete
        this role will result in a 400 status code.

        The system must have at least one "default" role. Attempting to delete
        the "default" role will result in a 400 status code.

        Users that have the role will be moved to the default role. If the
        role to be deleted had the `ASSIGNABLE_TO_CASE` permission and the
        "default" does not:
          * cases assigned to users with the role will be assigned
          * users with the admin role's will be notified about the affected cases.
      responses:
        '200':
          description: Success Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextualMessage'
...
components:
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        default:
          type: boolean
        permissions:
          type: integer
        permission_codes:
          type: array
          description: Array of permission codes this role has.
          example: ["CONFIGURE_SYSTEM", "READ_ACCOUNT", "UPDATE_ACCOUNT"]
          items:
            type: string
      required:
        - id
        - name
        - default
        - permissions
    AddRole:
      type: object
      properties:
        name:
          type: string
        default:
          type: boolean
        permissions:
          type: integer
      required:
        - name
    UpdateRole:
      type: object
      properties:
        name:
          type: string
        default:
          type: boolean
        permissions:
          type: integer
...
    ContextualMessage:
      type: object
      properties:
        message:
          type: string
          description: A contextual error message.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT